+++ react/features/base/util/index.js
@@ -1,3 +1,2 @@
 export * from './loadScript';
 export * from './randomUtil';
-export * from './roomnameGenerator';
